@@ -1,28 +1,45 @@
 name: localization_generator
 description: A window desktop application to generate localization config from excel file into .json file and LocaleKey class
 
-version: 1.0.1
+version: 2.0.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.1
-  filesystem_picker: ^1.0.4
-  filepicker_windows: ^1.0.0
-  excel: ^1.1.4
-  jin_widget_helper: ^2.0.1
-  toast: ^0.1.5
-  shared_preferences: ^0.5.12+4
-  path_provider: ^1.6.25
-  clipboard: ^0.1.2+8
+  cupertino_icons: ^1.0.3
+  filesystem_picker: ^2.0.0-nullsafety.0
+  filepicker_windows: ^2.0.0
+  excel: ^2.0.0-null-safety-3
+  sura_flutter: ^0.2.0
+  shared_preferences: ^2.0.5
+  path_provider: ^2.0.1
+  clipboard: ^0.1.3
+  hive: ^2.0.4
+  hive_flutter: ^1.0.0
+
+  toast:
+    git:
+      url: https://github.com/appdev/FlutterToast.git
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
+  hive_generator: ^1.1.0
+  build_runner: ^2.0.3
+
 flutter:
   uses-material-design: true
+
+  fonts:
+    - family: Kantumruy
+      fonts:
+        - asset: assets/fonts/Kantumruy-Light.ttf
+        - asset: assets/fonts/Kantumruy-Regular.ttf
+          weight: 500
+        - asset: assets/fonts/Kantumruy-Bold.ttf
+          weight: 700
